requirements/local.txt
apiview==1.3.25
beautifulsoup4==4.8.1
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.6
django-bootstrap4==1.0.1
django-cors-headers==3.2.0
django-debug-toolbar==2.0
django-extensions==2.2.5
django-import-export==2.0.1
django-suit==0.2.28
djangorestframework==3.10.2
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==1.4.4
odfpy==1.4.0
openpyxl==3.0.3
Pillow==6.2.1
python-decouple==3.1
pytz==2019.3
PyYAML==5.3
six==1.13.0
soupsieve==1.9.4
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
